+++ Crescendo2024/src/main/deploy/pathplanner/paths/LoadPiecePath.path
@@ -3,25 +3,25 @@
   "waypoints": [
     {
       "anchor": {
-        "x": 1.9126811185953203,
-        "y": 2.376375279066348
+        "x": 14.680121872092089,
+        "y": 2.3343970858557817
       },
       "prevControl": null,
       "nextControl": {
-        "x": 1.8126811185953202,
-        "y": 2.376375279066348
+        "x": 14.722383698266158,
+        "y": 2.243766307152117
       },
       "isLocked": false,
       "linkedName": null
     },
     {
       "anchor": {
-        "x": 1.0691193312211191,
-        "y": 1.0904579202642115
+        "x": 15.5,
+        "y": 1.0
       },
       "prevControl": {
-        "x": 1.6968366991565602,
-        "y": 2.177696294321803
+        "x": 15.479446690861732,
+        "y": 1.0440767136863907
       },
       "nextControl": null,
       "isLocked": false,
