+++ Crescendo2024/src/main/deploy/pathplanner/paths/SpeakerPath.path
@@ -3,25 +3,25 @@
   "waypoints": [
     {
       "anchor": {
-        "x": 1.830382407627542,
-        "y": 6.707344943511942
+        "x": 2.9928516999891146,
+        "y": 5.493438956802725
       },
       "prevControl": null,
       "nextControl": {
-        "x": 1.830382407627542,
-        "y": 6.71763228238236
+        "x": 1.9928516999891164,
+        "y": 5.493438956802725
       },
       "isLocked": false,
       "linkedName": null
     },
     {
       "anchor": {
-        "x": 1.830382407627542,
-        "y": 7.69492947507578
+        "x": 1.7069343411869782,
+        "y": 5.493438956802725
       },
       "prevControl": {
-        "x": 1.830382407627542,
-        "y": 6.69492947507578
+        "x": 1.7165369284957381,
+        "y": 5.493438956802725
       },
       "nextControl": null,
       "isLocked": false,
@@ -39,7 +39,7 @@
   },
   "goalEndState": {
     "velocity": 0,
-    "rotation": -138.30186567443516,
+    "rotation": 0,
     "rotateFast": false
   },
   "reversed": false,
